requirements.txt
pillow>=10.0.0
pyinstaller>=5.0
